registrationsコントローラをusersコントローラに変更/auth.tsをsession.tsとuser.tsに分割/ディレクトリ構成を整理
