--- app/views/api/auth/registrations/update.json.jbuilder
@@ -1,3 +1,0 @@
-  json.user do
-    json.partial! 'api/auth/sessions/user', user: @user
-  end
