+++ app/views/api/payments/index.json.jbuilder
@@ -4,7 +4,7 @@ json.array! @payments_group_by_paid_at do |date, payments|
     json.array!(payments) do |payment|
       json.extract! payment, :id, :amount, :detail, :paid_at, :settled, :settled_at
       json.user do
-        json.partial! 'api/auth/sessions/user', user: payment.user
+        json.partial! 'api/users/user', user: payment.user
       end
     end
   end
