+++ app/views/api/payments/update.json.jbuilder
@@ -1,4 +1,4 @@
 json.extract! @payment, :id, :amount, :detail, :paid_at, :settled, :settled_at
 json.user do
-  json.partial! 'api/auth/sessions/user', user: @payment.user
+  json.partial! 'api/users/user', user: @payment.user
 end
