+++ app/views/api/sessions/index.json.jbuilder
@@ -1,6 +1,6 @@
 if @user
   json.user do
-    json.partial! 'api/auth/sessions/user', user: @user
+    json.partial! 'api/users/user', user: @user
   end
   json.is_existed true
 else
