+++ app/views/api/users/update.json.jbuilder
@@ -1,0 +1,3 @@
+  json.user do
+    json.partial! 'api/users/user', user: @user
+  end
