+++ config/routes.rb
@@ -1,11 +1,7 @@
 Rails.application.routes.draw do
   namespace :api, defaults: { format: :json } do
-    namespace :auth do
-      resources :registrations, only: %i[create, update]
-      resources :sessions, only: %i[index]
-      patch :registrations, to: 'registrations#update'
-    end
-
+    resources :users, only: %i[create update]
+    resources :sessions, only: %i[index]
     resources :payments
     resources :teams, only: %i[show] do
       patch :payments, to: 'teams/payments#update'
